Aprimorando a estrutura
