+++ 02-ignite-timer/src/pages/Home/components/NewCycleForm/index.tsx
@@ -1,0 +1,53 @@
+import { zodResolver } from '@hookform/resolvers/zod'
+import { useForm } from 'react-hook-form'
+import * as zod from 'zod'
+import { FormContainer, TaskInput, MinutesAmountInput } from './styles'
+
+export function NewCycleForm() {
+
+  const newCycleFormValidationSchema = zod.object({
+    task: zod.string().min(1, 'Informe a tarefa'),
+    minutesAmount: zod.number().min(5).max(60),
+  })
+
+  const { register, handleSubmit, watch } = useForm<NewCycleFormData>({
+    resolver: zodResolver(newCycleFormValidationSchema),
+    defaultValues: {
+      task: '',
+      minutesAmount: 0,
+    },
+  })
+
+  
+
+  return (
+    <FormContainer>
+      <label htmlFor="task">Vou trabalhar em</label>
+      <TaskInput
+        id="task"
+        placeholder="Dê um nome para o seu projeto"
+        list="task-sugestions"
+        disabled={!!activeCycle}
+        {...register('task')}
+      />
+
+      <datalist id="task-sugestions">
+        <option value="Trabalhar" />
+        <option value="Dormir" />
+      </datalist>
+
+      <label htmlFor="minutesAmount">Durante</label>
+      <MinutesAmountInput
+        type="number"
+        id="minutesAmount"
+        placeholder="00"
+        step={5}
+        min={5}
+        max={60}
+        disabled={!!activeCycle}
+        {...register('minutesAmount', { valueAsNumber: true })}
+      />
+      <span>minutos</span>
+    </FormContainer>
+  )
+}
